[EXAMPLES] Reorganized `emma-examples` into submodules.
Reorganized `emma-examples` into three submodules.

The `emma-examples-library` module defines basic library methods, compiler
integration tests and abstract functional integration tests for those.

The `emma-flink-examples` module wraps the examples in `emma.onFlink` macros
and defines a Runner class (with main method) as well as Flink-specific
subclasses for the functional integration tests.

The `emma-spark-examples` is symmertric to `emma-flink-examples` and defines
the same extensions for Spark using a corresponding `emma.onSpark` macro.
