+++ emma-examples/emma-examples-library/src/main/scala/org/emmalanguage/examples/graphs/model.scala
