+++ emma-examples/emma-examples-library/src/main/scala/org/emmalanguage/examples/imdb/model.scala
