+++ emma-examples/emma-examples-library/src/test/resources/clustering/kmeans/points.tsv
