+++ emma-examples/emma-examples-library/src/test/resources/graphs/belief-prop/variables
