+++ emma-examples/emma-examples-library/src/test/resources/graphs/graph-col/edges.tsv
