+++ emma-examples/emma-examples-library/src/test/resources/graphs/spmv/vector
