+++ emma-examples/emma-examples-library/src/test/scala/org/emmalanguage/compiler/integration/graphs/TransitiveClosureITCase.scala
