+++ emma-examples/emma-examples-library/src/test/scala/org/emmalanguage/compiler/integration/text/WordCountITCase.scala
