+++ emma-examples/emma-examples-library/src/test/scala/org/emmalanguage/examples/graphs/BaseTriangleCountSpec.scala
