+++ emma-examples/emma-examples-library/src/test/scala/org/emmalanguage/examples/ml/classification/BaseNaiveBayesSpec.scala
