+++ emma-examples/emma-examples-library/src/test/scala/org/emmalanguage/examples/text/BaseWordCountSpec.scala
